python/evaluator/requirements.txt
influxdb
pyyaml
slacker
